requirements.txt
# UV Transfer Tool - Dependencies
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
Pillow>=9.0.0
tqdm>=4.60.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Optional FBX backends (install at least one)
# fbx>=2020.0           # Official Autodesk FBX SDK
# pyfbx>=0.1.0          # Alternative FBX parser
# pyassimp>=4.1.0       # Assimp wrapper

# Development (optional)
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.0.0
